requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
Flask-Migrate==3.1.1
PyMySQL==1.0.2
Waitress==2.1.1
imagekitio==3.0.1
mysql-connector-python==8.0.33
requests==2.28.2
SQLAlchemy==1.4.36
mysql-connector-python==8.0.33
ipython==8.12.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
